% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celegans.R
\docType{data}
\name{celegans}
\alias{celegans}
\title{Features extracted from hairpins of Caenorhabditis elegans.}
\format{A data frame with 1000 rows and 29 columns. The first 28 columns are
numeric features used in [1]. The last column is a logical variable
indicating
if the stem-loop is a pre-miRNA or not.}
\usage{
celegans
}
\description{
Small dataset of features extracted from C. elegans hairpins. The full
dataset is contained in the zip file "experiment_scripts.zip" that can be
downloaded from:
}
\details{
http://sourceforge.net/projects/sourcesinc/files/mirnass/
}
\references{
[1] Gudyś, A., Szcześniak, M. W., Sikora, M., & Makałowska, I. (2013).
HuntMi: an efficient and taxon-specific approach in pre-miRNA
identification. BMC bioinformatics, 14(1), 1.
}
\keyword{datasets}
